[Refactor] rename folder
