+++ src/components/form/checkboxField/index.spec.tsx
@@ -1,0 +1,7 @@
+describe("Some Tests", () => {
+  it("is true", () => {
+    expect(true).to.be.true;
+  });
+});
+
+export {};
